Move parameter-loading into zcash_proofs
