+++ zcash_proofs/src/hashreader.rs
